backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
mediapipe==0.10.14
ultralytics==8.2.28
timm==1.0.3
torch==2.3.0
torchvision==0.18.0
aiofiles==23.2.1
